add: refactoring the backend
